refactor tests for truffle 3

+++ scripts/test.sh
@@ -1,7 +1,13 @@
 #! /bin/bash
 
-testrpc &
-trpc_pid=$!
+output=$(nc -z localhost 8545; echo $?)
+[ $output -eq "0" ] && trpc_running=true
+if [ ! $trpc_running ]; then
+  echo "Starting our own testrpc node instance"
+  testrpc > /dev/null &
+  trpc_pid=$!
+fi
 truffle test
-kill -9 $trpc_pid
-
+if [ ! $trpc_running ]; then
+  kill -9 $trpc_pid
+fi
